requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make compile-versions
#
aio-pika==5.2.4
aiofiles==0.4.0           # via msp-rfcommon-library, sanic
aiohttp==3.5.4
aiokafka==0.5.0           # via msp-rfcommon-library
aiopg==0.16.0
aioredis==1.2.0           # via msp-rfcommon-library
aiormq==1.2.2             # via aio-pika
aiozk==0.12.0             # via msp-rfcommon-library
amqp==2.4.1               # via kombu
arrow==0.13.0
async-timeout==3.0.1      # via aiohttp, aioredis, msp-rfcommon-library
attrs==18.2.0             # via aiohttp, msp-rfcommon-library
billiard==3.5.0.5         # via celery
bleach==3.1.0             # via msp-rfcommon-library, readme-renderer
celery==4.2.1
certifi==2018.11.29       # via msp-rfcommon-library, requests
chardet==3.0.4            # via aiohttp, msp-rfcommon-library, requests
docutils==0.14            # via msp-rfcommon-library, readme-renderer
hiredis==1.0.0            # via aioredis, msp-rfcommon-library
httptools==0.0.11         # via msp-rfcommon-library, sanic
idna==2.8                 # via msp-rfcommon-library, requests, yarl
kafka-python==1.4.3       # via aiokafka, msp-rfcommon-library
kombu==4.2.2.post1        # via celery
marshmallow==2.18.0
msp-rfcommon-library==1.0.56
multidict==4.5.2          # via aiohttp, msp-rfcommon-library, sanic, yarl
newrelic==4.16.1.117
pamqp==2.1.0              # via aiormq
pkginfo==1.5.0.1          # via msp-rfcommon-library, twine
psycopg2-binary==2.7.7    # via aiopg
psycopg2==2.7.7
pygments==2.3.1           # via msp-rfcommon-library, readme-renderer
python-dateutil==2.7.5
python-json-logger==0.1.10
pytz==2018.9              # via celery, msp-rfcommon-library
pyyaml==3.13
readme-renderer==24.0     # via msp-rfcommon-library, twine
requests-toolbelt==0.8.0  # via msp-rfcommon-library, twine
requests==2.21.0          # via msp-rfcommon-library, requests-toolbelt, twine
sanic-cors==0.9.7
sanic-plugins-framework==0.6.5  # via sanic-cors
sanic==18.12.0
six==1.12.0               # via bleach, msp-rfcommon-library, python-dateutil, readme-renderer
sqlalchemy==1.2.16
tqdm==4.29.1              # via msp-rfcommon-library, twine
twine==1.12.1             # via msp-rfcommon-library
ujson==1.35
urllib3==1.24.1           # via msp-rfcommon-library, requests
uvloop==0.12.0
vine==1.2.0               # via amqp
webencodings==0.5.1       # via bleach, msp-rfcommon-library
websockets==6.0           # via msp-rfcommon-library, sanic
yarl==1.3.0               # via aio-pika, aiohttp, aiormq, msp-rfcommon-library

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.4.0        # via twine
